Bugfix 样式调整

+++ app/src/main/java/com/github/log2c/b1lib1li_tv/model/FeedModel.java
@@ -6,7 +6,7 @@ public class FeedModel {
     private boolean has_more;
     private String offset;
     private String update_baseline;
-    private int update_num;
+    private long update_num;
     private List<ItemsModel> items;
 
     public boolean isHas_more() {
@@ -33,11 +33,11 @@ public class FeedModel {
         this.update_baseline = update_baseline;
     }
 
-    public int getUpdate_num() {
+    public long getUpdate_num() {
         return update_num;
     }
 
-    public void setUpdate_num(int update_num) {
+    public void setUpdate_num(long update_num) {
         this.update_num = update_num;
     }
 
@@ -98,7 +98,7 @@ public class FeedModel {
 
         public static class BasicModel {
             private String comment_id_str;
-            private int comment_type;
+            private long comment_type;
             private LikeIconModel like_icon;
             private String rid_str;
 
@@ -110,11 +110,11 @@ public class FeedModel {
                 this.comment_id_str = comment_id_str;
             }
 
-            public int getComment_type() {
+            public long getComment_type() {
                 return comment_type;
             }
 
-            public void setComment_type(int comment_type) {
+            public void setComment_type(long comment_type) {
                 this.comment_type = comment_type;
             }
 
@@ -137,7 +137,7 @@ public class FeedModel {
             public static class LikeIconModel {
                 private String action_url;
                 private String end_url;
-                private int id;
+                private long id;
                 private String start_url;
 
                 public String getAction_url() {
@@ -156,11 +156,11 @@ public class FeedModel {
                     this.end_url = end_url;
                 }
 
-                public int getId() {
+                public long getId() {
                     return id;
                 }
 
-                public void setId(int id) {
+                public void setId(long id) {
                     this.id = id;
                 }
 
@@ -219,14 +219,14 @@ public class FeedModel {
                 private boolean following;
                 private String jump_url;
                 private String label;
-                private int mid;
+                private long mid;
                 private String name;
                 private OfficialVerifyModel official_verify;
                 private PendantModel pendant;
                 private String pub_action;
                 private String pub_location_text;
                 private String pub_time;
-                private int pub_ts;
+                private long pub_ts;
                 private String type;
                 private VipModel vip;
 
@@ -278,11 +278,11 @@ public class FeedModel {
                     this.label = label;
                 }
 
-                public int getMid() {
+                public long getMid() {
                     return mid;
                 }
 
-                public void setMid(int mid) {
+                public void setMid(long mid) {
                     this.mid = mid;
                 }
 
@@ -334,11 +334,11 @@ public class FeedModel {
                     this.pub_time = pub_time;
                 }
 
-                public int getPub_ts() {
+                public long getPub_ts() {
                     return pub_ts;
                 }
 
-                public void setPub_ts(int pub_ts) {
+                public void setPub_ts(long pub_ts) {
                     this.pub_ts = pub_ts;
                 }
 
@@ -361,10 +361,10 @@ public class FeedModel {
                 public static class DecorateModel {
                     private String card_url;
                     private FanModel fan;
-                    private int id;
+                    private long id;
                     private String jump_url;
                     private String name;
-                    private int type;
+                    private long type;
 
                     public String getCard_url() {
                         return card_url;
@@ -382,11 +382,11 @@ public class FeedModel {
                         this.fan = fan;
                     }
 
-                    public int getId() {
+                    public long getId() {
                         return id;
                     }
 
-                    public void setId(int id) {
+                    public void setId(long id) {
                         this.id = id;
                     }
 
@@ -406,11 +406,11 @@ public class FeedModel {
                         this.name = name;
                     }
 
-                    public int getType() {
+                    public long getType() {
                         return type;
                     }
 
-                    public void setType(int type) {
+                    public void setType(long type) {
                         this.type = type;
                     }
 
@@ -418,7 +418,7 @@ public class FeedModel {
                         private String color;
                         private boolean is_fan;
                         private String num_str;
-                        private int number;
+                        private long number;
 
                         public String getColor() {
                             return color;
@@ -444,11 +444,11 @@ public class FeedModel {
                             this.num_str = num_str;
                         }
 
-                        public int getNumber() {
+                        public long getNumber() {
                             return number;
                         }
 
-                        public void setNumber(int number) {
+                        public void setNumber(long number) {
                             this.number = number;
                         }
                     }
@@ -456,7 +456,7 @@ public class FeedModel {
 
                 public static class OfficialVerifyModel {
                     private String desc;
-                    private int type;
+                    private long type;
 
                     public String getDesc() {
                         return desc;
@@ -466,28 +466,28 @@ public class FeedModel {
                         this.desc = desc;
                     }
 
-                    public int getType() {
+                    public long getType() {
                         return type;
                     }
 
-                    public void setType(int type) {
+                    public void setType(long type) {
                         this.type = type;
                     }
                 }
 
                 public static class PendantModel {
-                    private int expire;
+                    private long expire;
                     private String image;
                     private String image_enhance;
                     private String image_enhance_frame;
                     private String name;
-                    private int pid;
+                    private long pid;
 
-                    public int getExpire() {
+                    public long getExpire() {
                         return expire;
                     }
 
-                    public void setExpire(int expire) {
+                    public void setExpire(long expire) {
                         this.expire = expire;
                     }
 
@@ -523,30 +523,30 @@ public class FeedModel {
                         this.name = name;
                     }
 
-                    public int getPid() {
+                    public long getPid() {
                         return pid;
                     }
 
-                    public void setPid(int pid) {
+                    public void setPid(long pid) {
                         this.pid = pid;
                     }
                 }
 
                 public static class VipModel {
-                    private int avatar_subscript;
+                    private long avatar_subscript;
                     private String avatar_subscript_url;
                     private long due_date;
                     private LabelModel label;
                     private String nickname_color;
-                    private int status;
-                    private int theme_type;
-                    private int type;
+                    private long status;
+                    private long theme_type;
+                    private long type;
 
-                    public int getAvatar_subscript() {
+                    public long getAvatar_subscript() {
                         return avatar_subscript;
                     }
 
-                    public void setAvatar_subscript(int avatar_subscript) {
+                    public void setAvatar_subscript(long avatar_subscript) {
                         this.avatar_subscript = avatar_subscript;
                     }
 
@@ -582,33 +582,33 @@ public class FeedModel {
                         this.nickname_color = nickname_color;
                     }
 
-                    public int getStatus() {
+                    public long getStatus() {
                         return status;
                     }
 
-                    public void setStatus(int status) {
+                    public void setStatus(long status) {
                         this.status = status;
                     }
 
-                    public int getTheme_type() {
+                    public long getTheme_type() {
                         return theme_type;
                     }
 
-                    public void setTheme_type(int theme_type) {
+                    public void setTheme_type(long theme_type) {
                         this.theme_type = theme_type;
                     }
 
-                    public int getType() {
+                    public long getType() {
                         return type;
                     }
 
-                    public void setType(int type) {
+                    public void setType(long type) {
                         this.type = type;
                     }
 
                     public static class LabelModel {
                         private String bg_color;
-                        private int bg_style;
+                        private long bg_style;
                         private String border_color;
                         private String img_label_uri_hans;
                         private String img_label_uri_hans_static;
@@ -628,11 +628,11 @@ public class FeedModel {
                             this.bg_color = bg_color;
                         }
 
-                        public int getBg_style() {
+                        public long getBg_style() {
                             return bg_style;
                         }
 
-                        public void setBg_style(int bg_style) {
+                        public void setBg_style(long bg_style) {
                             this.bg_style = bg_style;
                         }
 
@@ -834,12 +834,12 @@ public class FeedModel {
                         private String bvid;
                         private String cover;
                         private String desc;
-                        private int disable_preview;
+                        private long disable_preview;
                         private String duration_text;
                         private String jump_url;
                         private StatModel stat;
                         private String title;
-                        private int type;
+                        private long type;
 
                         public String getAid() {
                             return aid;
@@ -881,11 +881,11 @@ public class FeedModel {
                             this.desc = desc;
                         }
 
-                        public int getDisable_preview() {
+                        public long getDisable_preview() {
                             return disable_preview;
                         }
 
-                        public void setDisable_preview(int disable_preview) {
+                        public void setDisable_preview(long disable_preview) {
                             this.disable_preview = disable_preview;
                         }
 
@@ -921,11 +921,11 @@ public class FeedModel {
                             this.title = title;
                         }
 
-                        public int getType() {
+                        public long getType() {
                             return type;
                         }
 
-                        public void setType(int type) {
+                        public void setType(long type) {
                             this.type = type;
                         }
 
@@ -983,15 +983,15 @@ public class FeedModel {
                 }
 
                 public static class TopicModel {
-                    private int id;
+                    private long id;
                     private String jump_url;
                     private String name;
 
-                    public int getId() {
+                    public long getId() {
                         return id;
                     }
 
-                    public void setId(int id) {
+                    public void setId(long id) {
                         this.id = id;
                     }
 
@@ -1076,14 +1076,14 @@ public class FeedModel {
                 }
 
                 public static class CommentModel {
-                    private int count;
+                    private long count;
                     private boolean forbidden;
 
-                    public int getCount() {
+                    public long getCount() {
                         return count;
                     }
 
-                    public void setCount(int count) {
+                    public void setCount(long count) {
                         this.count = count;
                     }
 
@@ -1097,14 +1097,14 @@ public class FeedModel {
                 }
 
                 public static class ForwardModel {
-                    private int count;
+                    private long count;
                     private boolean forbidden;
 
-                    public int getCount() {
+                    public long getCount() {
                         return count;
                     }
 
-                    public void setCount(int count) {
+                    public void setCount(long count) {
                         this.count = count;
                     }
 
@@ -1118,15 +1118,15 @@ public class FeedModel {
                 }
 
                 public static class LikeModel {
-                    private int count;
+                    private long count;
                     private boolean forbidden;
                     private boolean status;
 
-                    public int getCount() {
+                    public long getCount() {
                         return count;
                     }
 
-                    public void setCount(int count) {
+                    public void setCount(long count) {
                         this.count = count;
                     }
